change datatype

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -10,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	int cnt = 0;
+	size_t cnt = 0;
 
 	while (h)
 	{
@@ -21,4 +21,3 @@ size_t print_listint(const listint_t *h)
 
 	return (cnt);
 }
-
